% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lfdr}
\alias{lfdr}
\title{Estimate local False Discovery Rate (FDR)}
\usage{
lfdr(p, pi0 = NULL, trunc = TRUE, monotone = TRUE, transf = c("probit",
  "logit"), adj = 1.5, eps = 10^-8, ...)
}
\arguments{
\item{p}{A vector of p-values (only necessary input).}

\item{pi0}{Estimated proportion of true null p-values. If NULL, then \code{\link{pi0est}} is called.}

\item{trunc}{If TRUE, local FDR values >1 are set to 1. Default is TRUE.}

\item{monotone}{If TRUE, local FDR values are non-decreasing with increasing p-values. Default is TRUE; this is recommended.}

\item{transf}{Either a "probit" or "logit" transformation is applied to the p-values so that a local FDR estimate can be formed that
does not involve edge effects of the [0,1] interval in which the p-values lie.}

\item{adj}{Numeric value that is applied as a multiple of the smoothing bandwidth used in the density estimation. Default is \code{adj=1.0}.}

\item{eps}{Numeric value that is threshold for the tails of the empirical p-value distribution. Default is 10^-8.}

\item{\ldots}{Additional arguments, passed to \code{\link{pi0est}}.}
}
\value{
A vector of estimated local FDR values, with each entry corresponding to the entries of the input p-value vector \code{p}.
}
\description{
Estimate the local FDR values from p-values.
}
\details{
The local FDR is defined to be the FDR in a small neighborhood about the value of a test statistic.  This is
typically formulated in a Bayesian framework to multiple testing.  Technical details can be found in Efron et al. (2001)
and Storey (2003) in the references below.  The following is a brief, intuitive summary.  Suppose that each hypothesis
test is a true null hypothesis with probability \eqn{\pi_0}{pi_0}. Suppose that the probability density function of a
true null hypothesis test is \eqn{f_0}, whereas for a true alternative it is \eqn{f_1}.  The marginal probability
density function can be written as \eqn{f = \pi_0 f_0 + (1-\pi_0) f_1}{f = pi_0*f_0 + (1-pi_0)*f_1}.  The local FDR for
a given p-value \eqn{p} is then defined to be:
\deqn{\frac{\pi_0 f_0(p)}{f(p)}.}{pi_0 * f_0(p) / f(p).}
For p-values, the assumption is that p-values derived from true null hypotheses are distributed Uniform(0,1),
in which case \eqn{f_0(p)=1}.  Also, \eqn{\pi_0}{pi_0} may be estimated from the p-values via the function \code{\link{pi0est}}.
This leaves estimation of \eqn{f(p)}, which is the density estimated from the set of observed p-values.

The strategy taken in \code{\link{lfdr}} is that an estimate of the density of the p-values \eqn{f(p)} is formed by utilizing the \code{density} function provided by the \code{\link{stats}} package. This allows one to estimate local FDRs very generally, without being limited to the two-sample t-tests which are of primary focus in Efron (2001) and references that follow.  In those papers, specialized estimators are formulated that rely substantially on the test-statistic used.  Here, one may apply \code{\link{lfdr}} as long as the Uniform(0,1) distribution is a reasonable assumption for the p-values coming from true null hypotheses.
}
\examples{
# import data
data(hedenfalk)
p <- hedenfalk$p
lfdrVals <- lfdr(p)

# plot local FDR values
qobj = qvalue(p)
hist(qobj)
}
\author{
John D. Storey \email{jstorey@princeton.edu}, Andrew J. Bass
}
\references{
Efron B, Tibshirani R, Storey JD, and Tisher V. (2001) Empirical Bayes analysis
of a microarray experiment. Journal of the American Statistical Association, 96: 1151-1160. \cr
\url{http://www.tandfonline.com/doi/abs/10.1198/016214501753382129}

Storey JD. (2003) The positive false discovery rate: A Bayesian
interpretation and the q-value. Annals of Statistics, 31: 2013-2035. \cr
\url{http://projecteuclid.org/DPubS/Repository/1.0/Disseminate?view=body&id=pdf_1&handle=euclid.aos/1074290335}

Storey JD. (2011) False discovery rates. In \emph{International Encyclopedia of Statistical Science}. \cr
\url{http://genomine.org/papers/Storey_FDR_2011.pdf} \cr
\url{http://www.springer.com/statistics/book/978-3-642-04897-5}
}
\seealso{
\code{\link{qvalue}}, \code{\link{pi0est}}, \code{\link{hist.qvalue}}
}
\keyword{Discovery}
\keyword{False}
\keyword{Rate,}
\keyword{lfdr}
\keyword{local}

