\name{qvalue.gui}
\alias{qvalue.gui}
\title{A graphical user interface for QValue}
\description{
  A graphical user interface for QValue
}
\usage{
qvalue.gui(dummy = NULL)
}
\arguments{
  \item{dummy}{Dummy argument to facilitate documentation (no arguments required)}
}
\details{
  This is a point-and-click version of \code{qvalue}.  See the manual at \url{http://faculty.washington.edu/~jstorey/qvalue/manual.pdf} for documentation.
}
\value{
  Nothing of interest.
}
\author{Alan R. Dabney \email{adabney@u.washington.edu}}
\seealso{\code{\link{qvalue}}, \code{\link{qplot}}, \code{\link{qwrite}}, \code{\link{qsummary}}}
\examples{
\dontrun{
qvalue.gui()
}
}
\keyword{misc}
