% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvalue_truncp.R
\name{qvalue_truncp}
\alias{qvalue_truncp}
\title{Q-value on truncated p-values}
\usage{
qvalue_truncp(
  p,
  fdr.level = NULL,
  pfdr = FALSE,
  lfdr.out = TRUE,
  pi0 = NULL,
  ...
)
}
\description{
Calculate \code{\link{qvalue}}, but with p-values re-scaled to be in `[0, 1]` range
}
\seealso{
\code{\link{qvalue}}
}
