% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qwrite}
\alias{qwrite}
\title{Write results to file}
\usage{
qwrite(qobj, filename = "my-qvalue-results.txt", ...)
}
\arguments{
\item{qobj}{A q-value object.}

\item{filename}{Output filename (optional).}

\item{\dots}{Additional arguments; currently unused.}
}
\value{
Nothing of interest.
}
\description{
Write the results of the q-value object to a file.
}
\details{
The output file lists the estimate of \eqn{\pi_0}{pi_0}, which is the
proportion of true null hypotheses. It also lists each p-value and
corresponding q-values and local FDR values, one per line. If an FDR significance level was
specified in the call to \code{\link{qvalue}}, the significance level
is printed below the estimate of \eqn{\pi_0}{pi_0}, and an indicator
of significance is included as a fourth column for each p-value and q-value.
}
\examples{
# import data
data(hedenfalk)
p <- hedenfalk$p

# write q-value object
qwrite(qobj, file="myresults.txt")
}
\author{
John D. Storey \email{jstorey@princeton.edu}, Andrew J. Bass
}
\seealso{
\code{\link{qvalue}}, \code{\link{plot.qvalue}}, \code{\link{summary.qvalue}}
}
\keyword{qwrite}
\keyword{write,}

